+++ app/users/dashboard/organizations/[id]/page.tsx
@@ -2,8 +2,8 @@
 
 import { notFound } from 'next/navigation';
 import { useState, use } from 'react';
-import { Button } from '@/components/ui/button';
-import { Badge } from '@/components/ui/badge';
+import { Button } from '@/components/shadcn/button';
+import { Badge } from '@/components/shadcn/badge';
 import { Department } from '@/types/employee/departments';
 import {
   Building2,
